requirements.txt
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.2.0
colorama==0.4.6
comtypes==1.4.9
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
daphne==4.1.2
Django==4.2.9
django-admin-datta==1.0.15
django-admin-rangefilter==0.13.2
django-api-generator==1.0.17
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-dynamic-datatb==1.0.25
django-filter==24.3
django-pwa==2.0.1
djangorestframework==3.15.2
et_xmlfile==2.0.0
fonttools==4.56.0
fpdf==1.7.2
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywifi==1.1.12
qrcode==8.0
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
str2bool==1.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.7.0
zope.interface==7.2